Break rayFunctions into separate files
