+++ src/rayFunctions/rayAlloc.ts
@@ -1,0 +1,9 @@
+import { IRay } from "../types";
+
+class Ray implements IRay {
+  constructor(public x0 = NaN, public y0 = NaN, public dirX = NaN, public dirY = NaN) {}
+}
+
+export function rayAlloc(): IRay {
+  return new Ray();
+}
