+++ src/rayFunctions/rayPointAt.ts
@@ -1,0 +1,7 @@
+import { IRay } from "../types";
+import { vecAlloc } from "../vecFunctions/vecAlloc";
+import { vecReset } from "../vecFunctions/vecReset";
+
+export function rayPointAt(ray: IRay, t: number, out = vecAlloc()) {
+  return vecReset(ray.x0 + ray.dirX * t, ray.y0 + ray.dirY * t, out);
+}
