+++ src/rayFunctions/rayReset.ts
@@ -1,0 +1,9 @@
+import { rayAlloc } from "./rayAlloc";
+
+export function rayReset(x0: number, y0: number, dirX: number, dirY: number, out = rayAlloc()) {
+  out.x0 = x0;
+  out.y0 = y0;
+  out.dirX = dirX;
+  out.dirY = dirY;
+  return out;
+}
